change address type in get_new_stack_top() from unsigned int to uint32_t

+++ user/libthread/thr_lib_helper.c
@@ -1,16 +1,17 @@
-#include "thr_lib_helper.h"
+#include <stdint.h>
+#include <thr_lib_helper.h>
 
 /**
  * @brief Root thread stack low
  */
-static unsigned int root_thread_stack_low;
+static uint32_t root_thread_stack_low;
 
 /** @brief Get current root thread stack low
  *  
  *  @return Current root thread stack low
  *  @bug Will eventually put this call in libautostack
  */
-unsigned int get_root_thread_stack_low() {
+uint32_t get_root_thread_stack_low() {
 
     return 0xffffe000;
 }
@@ -33,7 +34,7 @@ unsigned int get_root_thread_stack_low() {
  *  collection.
  *
  */
-unsigned int get_new_stack_top(int count, 
+uint32_t get_new_stack_top(int count, 
         unsigned int stack_size) {
 
     // Once we create a new thread, the root thread's stack region is
@@ -43,15 +44,15 @@ unsigned int get_new_stack_top(int count,
     }
 
     // Assume pages on the stack grow down continuouly
-    unsigned int new_page_base = (root_thread_stack_low - 
+    uint32_t new_page_base = (root_thread_stack_low - 
             count * stack_size) & PAGE_ALIGN_MASK;
 
-    unsigned int old_page_base = (root_thread_stack_low - 
+    uint32_t old_page_base = (root_thread_stack_low - 
             (count - 1) * stack_size) & PAGE_ALIGN_MASK;
 
     // Allocate if previously allocated pages are not enough
     if(new_page_base != old_page_base) {
-        lprintf("new_page_base: %x", new_page_base);
+        lprintf("new_page_base: %x", (unsigned int)new_page_base);
         if(new_pages((void *)new_page_base, 
                     old_page_base - new_page_base) != 0) {
             lprintf("new_pages fail");
@@ -66,7 +67,7 @@ unsigned int get_new_stack_top(int count,
 
     // The 1st available new stack position is last thread's stack low - 1
     // Keep decrementing until it aligns with 4
-    unsigned int new_stack_top = root_thread_stack_low - 
+    uint32_t new_stack_top = root_thread_stack_low - 
         (count - 1) * stack_size - 1;
     while(new_stack_top % ALIGNMENT != 0) {
         new_stack_top--;
